requirements.txt
# AEELR Requirements
# Attention-Enhanced EfficientNet with Label Refinement

# Core deep learning
tensorflow>=2.13.0
keras>=2.13.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
scipy>=1.11.0
Pillow>=10.0.0

# Label refinement
cleanlab>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Demo
gradio>=4.0.0

# Utilities
tqdm>=4.65.0

# Optional: For better performance
# tensorflow-gpu>=2.13.0  # If using GPU
